Refactor and comment more
