+++ lib/wnew_csv2json.js
@@ -1,8 +1,8 @@
+'use strict'
 /**
  * Gulp plugin split a csv file into json files.
  * @file
  */
-'use strict'
 
 const through = require('through2')
 
@@ -10,9 +10,21 @@ const ROW_SEPARATOR = '\n'
 const COL_SEPARATOR = ','
 const MENU_FILENAME = 'menu.json'
 
+/**
+ * wnew csv2json
+ * @return {Stream} [description]
+ */
 module.exports = function() {
 
+  /**
+   * Transform
+   * @param  {Vinyl}   file      [description]
+   * @param  {String}   encode   [description]
+   * @param  {Function} callback [description]
+   * @return {void}              [description]
+   */
   function transform(file, encode, callback) {
+
     // get single file input, so far
     const cont = file.contents.toString()
 
@@ -70,6 +82,11 @@ module.exports = function() {
     callback()
   }
 
+  /**
+   * Flush
+   * @param  {Function} callback [description]
+   * @return {Void}              [description]
+   */
   function flush(callback) {
     callback()
   }
